# internal/endpoint/auth/handler_model.go
package auth

type LoginRequest struct {
	Username string `json:"username"`
	Password string `json:"password"`
}
type LoginResponse struct {
	Access string `json:"access"`
}

type RegisterRequest struct {
	Email    string `json:"email"`
	Username string `json:"username"`
	Password string `json:"password"`
}
type RegisterResponse struct {
	Access string `json:"access"`
}